Ordering refactoring
